requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.7
WTForms==3.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
Flask-Uploads==0.2.1
Flask-Mail==0.9.1
email_validator==2.0.0